+++ tests/test_flow_runs.py
@@ -1,0 +1,40 @@
+import time
+
+import pytest
+
+import prefect.client.schemas as client_schemas
+from prefect import flow
+from prefect.exceptions import FlowWaitTimeout
+from prefect.flow_runs import wait_for_flow_run
+from prefect.states import Completed
+
+
+async def test_create_then_wait_for_flow_run(prefect_client):
+    @flow
+    def foo():
+        pass
+
+    flow_run = await prefect_client.create_flow_run(
+        foo, name="petes-flow-run", state=Completed()
+    )
+    assert isinstance(flow_run, client_schemas.FlowRun)
+
+    lookup = await wait_for_flow_run(prefect_client, flow_run.id, poll_interval=0)
+    # Estimates will not be equal since time has passed
+    assert lookup == flow_run
+    assert flow_run.state.is_final()
+
+
+async def test_create_then_wait_timeout(prefect_client):
+    @flow
+    def foo():
+        time.sleep(9999)
+
+    flow_run = await prefect_client.create_flow_run(
+        foo,
+        name="petes-flow-run",
+    )
+    assert isinstance(flow_run, client_schemas.FlowRun)
+
+    with pytest.raises(FlowWaitTimeout):
+        await wait_for_flow_run(prefect_client, flow_run.id, timeout=0)
